static analysis: separate parser code into separate subdirectory
