+++ internal/staticanalysis/obfuscation/detection/generate_signals_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 )
 
-const jsParserPath = "../parsing/babel-parser.js"
+const jsParserPath = "../../parsing/js/babel-parser.js"
 
 func symbolEntropySummary(symbols []string) stats.SampleStatistics {
 	probs := stringentropy.CharacterProbabilities(symbols)
