+++ internal/staticanalysis/parsing/js/babel-parser.js
