+++ internal/staticanalysis/parsing/js/js_parsing.go
@@ -1,4 +1,4 @@
-package parsing
+package js
 
 import (
 	"encoding/json"
