+++ internal/staticanalysis/parsing/js/parsing_types.go
@@ -1,4 +1,4 @@
-package parsing
+package js
 
 import (
 	"fmt"
